+++ src/jmbde/Pages/DeviceTypes/Edit.cshtml.cs
@@ -82,7 +82,7 @@ namespace JMuelbert.BDE.Pages.DeviceTypes {
         /// </summary>
         /// <returns>The get async.</returns>
         /// <param name="id">Identifier.</param>
-        public async Task<IActionResult> OnGetAsync (long? id) {
+        public async Task<IActionResult> OnGetAsync (int? id) {
             _logger.LogDebug ($"DeviceTypes/Edit/OnGetAsync({ id })");
             if (id == null) {
                 return NotFound ();
@@ -101,7 +101,7 @@ namespace JMuelbert.BDE.Pages.DeviceTypes {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<IActionResult> OnPostAsync (long? id) {
+        public async Task<IActionResult> OnPostAsync (int? id) {
             _logger.LogDebug ($"DeviceTypes/Edit/OnPostAsync({ id })");
             if (!ModelState.IsValid) {
                 return Page ();
